Login-index
I made some changes on the login page. now we have a label where we receive the generated password. I add the footer  on index page and also the connection between login and index page previously validate the credentials. It didn't show a error message and i still working on it.